OMG, I'm a genius!
Finally fixed the issues we have had with images. Some grayscale images
were inverted and some color images could be grayscale. This was caused
by the fact that those images had a palette. That palette was not being
applied when the depth was being read as 8-bits (indexed).

BOOM BABY!